Just playing around (bypassing NCURSE lib)
Not essential for project but can be used as a template/for ideas. Pulls text in active time and has the safety guards to "fail safe" --> Will time out and return to accepting normal keyboard input if enough time passes, or if CTRL+C is pressed. (Or other control sequences --> Effectively leaving some "untouched" keys).